+++ node/tests/resolvers.test.ts
@@ -30,6 +30,7 @@ describe('Graphql resolvers', () => {
       },
       {
         clients,
+        vtex,
       }
     )