+++ node/tests/vtexContext.ts
@@ -1,5 +1,4 @@
 import type { IOContext } from '@vtex/api'
-import { Segment } from '@vtex/api'
 
 import Service from '../index'
 import { Search } from '../clients/search'
@@ -14,10 +13,9 @@ const vtexContext = () => {
   const context = {} as IOContext
   const clients = {
     search: new Search(context),
-    segment: new Segment(context),
   }
 
-  const vtex = {} as IOContext
+  const vtex = {} as CustomIOContext
 
   return {
     context,